==> 1_model_runs/wflow_sbm/06_create_evaluation_config_files.ipynb <==
# Create wflow evaluation period config file using calibrated parameters
## Based on best performing ksathorfrac values

from glob import glob
from pathlib import Path

import toml
import pandas as pd
import xarray as xr

# Set Paths

# Set Paths
ROOT = Path('/gpfs/work1/0/wtrcycle/users/jaerts/model_refinement_pub/')
MODELS = Path(f'{ROOT}/model_parameters/wflow_sbm/')
AUXDATA = Path(f"{ROOT}/aux_data")
RESULTS = Path(f"{ROOT}/results/streamflow_evaluation/wflow_sbm_calibration/")

template = f"{AUXDATA}/wflow_sbm/template_wflow_sbm_evaluation.toml"

# Load calibration results

# Load calibration file
calibration_file = f"{RESULTS}/wflow_calibration_objective_function_overview.csv"
df = pd.read_csv(calibration_file, index_col='basin_id')

basin_ids = df.index.to_list()

len(basin_ids)

# Create evaluation period config files based on calibrated parameters

# Evaluation Period
for basin_id in basin_ids:
    calibration_values = int(df.loc[int(basin_id)].ksathorfrac)  
    
    config_file = toml.load(template)
    config_file['loglevel'] = "debug"      
    config_file['path_log'] = f"log_evaluation.txt"    
    
    # Add timeperiod
    config_file['starttime'] = '2008-01-01T00:00:00'
    config_file['endtime'] = '2017-12-31T00:00:00'
    
    # Add file location
    config_file['input']['path_forcing'] = f'ceh-gear_chess_camels-gb_{basin_id}_2000_2017.nc'
    config_file['input']['path_static'] = f'staticmaps_ksathorfrac_{calibration_values}.nc'

    # Add output
    config_file['state']['path_output'] = f'evaluation/outstate/outstates.nc'
    config_file['output']['path'] = f'evaluation/output.nc'
    config_file['csv']['path'] = f'evaluation/output.csv'
    
    # Add interception
    config_file['output']['vertical']['interception'] = 'interception'
    
    
   
    with open(f'{MODELS}/{basin_id}/wflow_sbm_evaluation.toml', 'w') as f:
        toml.dump(config_file, f)

# Create evaluation period parameter file

for basin_id in basin_ids:
    calibration_value = int(df.loc[int(basin_id)].ksathorfrac) 
    
    # Load dataset
    ds = xr.open_dataset(f'{MODELS}/{basin_id}/staticmaps.nc')
    # Multiply with calibration value
    ds['KsatHorFrac'] = ds.KsatHorFrac * (calibration_value / 100)

    # Save to netcdf
    fname = f'{MODELS}/{basin_id}/staticmaps_ksathorfrac_{calibration_value}.nc'

    write_job = ds.to_netcdf(fname, compute=False)
    with ProgressBar():
        write_job.compute()
